requirements.txt
pandas-datareader==0.9.0
pmdarima==1.7.0
python-dotenv==0.14.0
seaborn==0.10.1
tqdm==4.48.2
